refactor(test-harness): one Postgres harness, in scheduler-metadata
metadata-service had its own copy: a second container, its own per-test
`CREATE DATABASE`, and the same never-dropped databases. It also claimed the
`watchdog` feature reaped its container at process exit, which it does not — the
container outlived its test binary by as long as the rest of the suite took,
holding 20 databases. `pg_harness` had already hit that and answered it with an
explicit `#[dtor]`; the copy never got the fix.

Reuse was not possible in place. `network-scheduler` and `metadata-service` are
siblings, so metadata-service would have had to dev-depend on the whole
scheduler application to reach the harness, and the only entry point `pg-testkit`
exposes hands back a `PostgresStorage` holding the scheduler's advisory lock —
where metadata-service needs a URL it can build its own pools on. So the harness
moves down to `scheduler-metadata`, the leaf both already depend on and whose
MIGRATOR both already run, behind a `pg-testkit` feature that only
dev-dependencies turn on. `explain` moves with it (re-exported from
`postgres::explain`, where its callers already look).

The shared harness hands out URLs plus the `CaseDb` that owns them, so each
crate can tie the database to whatever value outlives its connections:
`PostgresStorage` for the scheduler, `TestApp` for the service. Its entry points
now run their `block_on` on a scratch thread, because `Runtime::block_on` panics
inside a runtime and the metadata-service suite is `#[tokio::test]`; a panic
there is resumed on the caller so the harness's own message survives the hop.

Full suite at default settings: 0 failed, peak PGDATA 224 MB / 13 live
databases, and no container left behind by either crate.
