+++ Cargo.toml
@@ -18,7 +18,7 @@ mvcc-chunks = ["dep:sqlx", "dep:bit-vec", "dep:humantime", "dep:scheduler-metada
 # Exposes the Postgres testcontainer harness (`scheduler_storage::test_harness::pg_harness`) to other
 # crates — e.g. reshuffle-sim. Implies the Postgres backend. The crate's own tests get the harness via
 # `cfg(test)` + dev-dependencies, so they don't need this feature.
-pg-testkit = ["mvcc-chunks", "dep:dtor", "dep:testcontainers", "dep:testcontainers-modules"]
+pg-testkit = ["mvcc-chunks", "scheduler-metadata/pg-testkit"]
 
 [dependencies]
 anyhow = "1.0.98"
@@ -60,21 +60,16 @@ tracing = "0.1.41"
 tracing-subscriber = { version = "0.3.19", features = ["env-filter", "parking_lot"] }
 secrecy = { version = "0.10.3", features = ["serde"] }
 sqlx = { workspace = true, optional = true }
-# Also dev-dependencies (for the crate's own tests); optional here so the `pg-testkit` feature can hand
-# the harness to downstream crates without forcing testcontainers into a plain build.
-dtor = { version = "1.0.5", optional = true }
-testcontainers = { version = "0.27.2", features = ["watchdog"], optional = true }
-testcontainers-modules = { version = "0.15", features = ["postgres"], optional = true }
 rustc-hash = "2"
 
 [dev-dependencies]
 claims = "0.8"
-dtor = "1.0.5"
 maplit = "1.0.2"
+# The crate's own tests get the harness here rather than through `pg-testkit`, which exists only to
+# hand it to downstream crates.
+scheduler-metadata = { path = "crates/scheduler-metadata", features = ["pg-testkit"] }
 # The iterative state-machine runner — whose generator can read the live SUT — lives in
 # `proptest-state-machine::iterative_runner`. Pinned to the HEAD of define-null/proptest's
 # `defnull/iterative-runner-2` branch. Both crates share the rev so they use one `proptest` build.
 proptest = { git = "https://github.com/define-null/proptest", rev = "253a26e9a6e4ac7c75d6265b441b1be00cf1e2c6" }
 proptest-state-machine = { git = "https://github.com/define-null/proptest", rev = "253a26e9a6e4ac7c75d6265b441b1be00cf1e2c6" }
-testcontainers = { version = "0.27.2", features = ["watchdog"] }
-testcontainers-modules = { version = "0.15", features = ["postgres"] }
