+++ crates/metadata-service/Cargo.toml
@@ -34,6 +34,6 @@ sha2 = "0.10"
 tower-http = { version = "0.6", features = ["timeout"] }
 
 [dev-dependencies]
-testcontainers = { version = "0.27.2", features = ["watchdog"] }
-testcontainers-modules = { version = "0.15", features = ["postgres"] }
+# The shared Postgres harness; a dev-dependency, so no production build sees testcontainers.
+scheduler-metadata = { path = "../scheduler-metadata", features = ["pg-testkit"] }
 tower = { version = "0.5", features = ["util"] }   # ServiceExt::oneshot
